+++ front/js/confirmation.js
@@ -1,7 +1,11 @@
-function confirm(){
+const orderId = localStorage.getItem("orderId");
+console.log(orderId)
+ if(orderId) {
     const idNode = document.getElementById("orderId");
     idNode.innerText = localStorage.getItem("orderId");
-    console.log(localStorage.getItem("orderId"))
+    console.log(localStorage.getItem("orderId"));
     localStorage.clear();
+
+}else{
+  document.querySelector(".confirmation p").innerHTML = "Commande incorrecte"
 }
-confirm();
